requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-socketio>=5.10.0
python-engineio>=4.8.0
eventlet>=0.35.2
numpy>=1.26.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
jinja2>=3.1.2
aiofiles>=23.2.1
websockets>=12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
